abstract server setup
