+++ variables.env
@@ -1,3 +1,4 @@
+# Convert this into a .env file for local use
 MONGO_URI=mongo
 MOVIEDB_APIKEY=key
 MOVIEDB_BASEURL=https://api.themoviedb.org/3
